+++ experimental/packages/opentelemetry-exporter-metrics-otlp-http/src/platform/browser/OTLPMetricExporter.ts
@@ -38,8 +38,8 @@ class OTLPExporterBrowserProxy extends OTLPExporterBrowserBase<
 > {
   constructor(config?: OTLPMetricExporterOptions & OTLPExporterConfigBase) {
     super(config);
-    this._headers = Object.assign(
-      this._headers,
+    this.headers = Object.assign(
+      this.headers,
       baggageUtils.parseKeyPairsIntoRecord(
         getEnv().OTEL_EXPORTER_OTLP_METRICS_HEADERS
       )